Refactore home view components (#5).
