+++ src/app/app.routing.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { CarDetailViewComponent } from './components/car-detail-view/car-detail-view.component';
 import { CarListViewComponent } from './components/car-list-view/car-list-view.component';
-import { HomeViewComponent } from './components/home-view/home-view.component';
+import { HomeViewComponent } from './views/home-view/home-view.component';
 
 const routes: Routes = [
   { path: ':user/:car', component: CarDetailViewComponent, pathMatch: 'full' },
