+++ src/app/components/login/login.component.html
@@ -1,11 +1,9 @@
-<h1>{{ 'Home.Title' | translate }}</h1>
-
 <form clrForm [formGroup]="form" clrLayout="vertical" class="full-width">
-  <p>{{ 'Home.Description' | translate }}</p>
+  <p>{{ 'Login.Description' | translate }}</p>
   <clr-input-container>
-    <input clrInput formControlName="userId" placeholder="{{ 'Home.UserIdPlaceholder' | translate }}" autocomplete="off" required />
+    <input clrInput formControlName="userId" placeholder="{{ 'Login.UserIdPlaceholder' | translate }}" autocomplete="off" required />
     <clr-control-error *clrIfError="'required'">{{ 'Validation.Required' | translate }}</clr-control-error>
     <clr-control-error *clrIfError="'pattern'">{{ 'Validation.OnlyInteger' | translate }}</clr-control-error>
   </clr-input-container>
-  <button class="btn btn-primary btn-block" type="button" (click)="submit()">{{ 'Home.LoginButton' | translate }}</button>
+  <button class="btn btn-primary btn-block" type="button" (click)="submit()">{{ 'Login.SubmitButtonText' | translate }}</button>
 </form>
