+++ src/app/views/home-view/home-view.component.html
@@ -1,0 +1,3 @@
+<h1>{{ 'Home.Title' | translate }}</h1>
+
+<myf-login></myf-login>
