+++ src/app/views/home-view/home-view.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'myf-home-view',
+  templateUrl: './home-view.component.html'
+})
+export class HomeViewComponent {}
